Split out OS-specific function.
